requirements.txt
snowflake-connector-python==3.12.3
python-dotenv==1.0.1
pandas==2.2.3
pytest==8.3.4
